Migrate Realtime Channel tests
